cp: Overwrite symlinks on another device
When target of a copy operation is a symlink, it is not the same file as
the source, so it is safe to overwrite it, regardless of if it is the
same device as the source or a different device.

Previous check only allowed overwrite on the same device and produced an
error when run across different file systems.